21/04/67

+++ server/models/incomeModel.js
@@ -3,7 +3,7 @@ const { DataTypes } = require("sequelize")
 module.exports = (sequelize,Sequelize) => {
     const Income = sequelize.define('Incomes', {
         incomeDetails: DataTypes.STRING,
-        amount: DataTypes.DECIMAL,
+        amount: DataTypes.INTEGER,
         incomeDate: DataTypes.DATE
     })
     return Income
